pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "slavoj"
dynamic = ["version"]
description = "A WhatsApp-based author conversation simulator using LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Lee Johnson", email = "leejavaa@gmail.com"},
]
dependencies = [
    "motor==3.7.0",
    "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["slavoj"]

[tool.setuptools.dynamic]
version = {attr = "slavoj.__version__"}